+++ gazebo/physics/ode/ODEJoint.cc
@@ -1266,7 +1266,8 @@ void ODEJoint::SetDamping(int _index, double _damping)
 {
   if (static_cast<unsigned int>(_index) < this->GetAngleCount())
   {
-    this->SetStiffnessDamping(_index, this->stiffnessCoefficient[_index],
+    this->SetStiffnessDamping(static_cast<unsigned int>(_index),
+      this->stiffnessCoefficient[_index],
       _damping);
   }
   else
@@ -1279,10 +1280,10 @@ void ODEJoint::SetDamping(int _index, double _damping)
 }
 
 //////////////////////////////////////////////////
-void ODEJoint::SetStiffnessDamping(int _index,
+void ODEJoint::SetStiffnessDamping(unsigned int _index,
   double _stiffness, double _damping, double _reference)
 {
-  if (_index < static_cast<int>(this->GetAngleCount()))
+  if (_index < this->GetAngleCount())
   {
     this->stiffnessCoefficient[_index] = _stiffness;
     this->dampingCoefficient[_index] = _damping;
